data_mining_cw1-master 3/requirements.txt
cycler==0.10.0
javabridge==1.0.18
kiwisolver==1.0.1
matplotlib==3.0.0
numpy==1.15.2
Pillow==5.3.0
pyparsing==2.2.2
python-dateutil==2.7.3
python-weka-wrapper3==0.1.5
six==1.11.0
